player upgrades for speed and strength

+++ Assets/TalonScripts/PlayerCombat.cs
@@ -13,7 +13,7 @@ public class PlayerCombat : MonoBehaviour
     public Transform attackPoint;
     public float attackRange = 1f;
     public LayerMask enemyLayers;
-    public float attackDamage = 1f;
+    public int attackDamage = 1;
 
     public int lifeStealAmount = 1;
     int timesHit = 0;
@@ -53,7 +53,7 @@ public class PlayerCombat : MonoBehaviour
                     }
                 }
                 enemies[0].GetComponent<EnemyKnockback>().Knockback(transform, knockbackForce);
-                enemies[0].GetComponent <EnemyHealth>().TakeDamage((int)attackDamage);  
+                enemies[0].GetComponent<EnemyHealth>().TakeDamage(attackDamage);  
                 Debug.Log("Hit " + enemies[0].name);
             }
